Ultimate le ultime cose, aggiunte le pagine per il login e per vedere le multe da lato amministratore

+++ api.php
@@ -1,6 +1,6 @@
 <?php 
 
-error_reporting(E_ALL);
+//error_reporting(E_ALL);
 
 function sendResponse($status = 200, $body = '', $conten_type = 'text/html')
 {
@@ -14,6 +14,7 @@ function sendResponse($status = 200, $body = '', $conten_type = 'text/html')
 }
 
 $autenticato = false;
+$serveAccesso = false;
 
 class serviceApi
 {
@@ -21,7 +22,6 @@ class serviceApi
     //metodi
     function __construct()
     {
-        
         $servername = "localhost";
         $username = "Multatore";
         $password = "Picasso69";
@@ -30,7 +30,7 @@ class serviceApi
 
         $options = [
             //PDO::ATTR_EMULATE_PREPARES   => false, // turn off emulation mode for "real" prepared statements
-            PDO::ATTR_ERRMODE            => PDO::ERRMODE_EXCEPTION, //turn on errors in the form of exceptions
+            //PDO::ATTR_ERRMODE            => PDO::ERRMODE_EXCEPTION, //turn on errors in the form of exceptions
            // PDO::ATTR_DEFAULT_FETCH_MODE => PDO::FETCH_ASSOC, //make the default fetch be an associative array
           ];
 
@@ -60,26 +60,34 @@ class serviceApi
         $foto = $_POST['foto'];
         $longitudine = $_POST['longitudine'];
         $latitudine = $_POST['latitudine'];
-        try
+
+        if(isset($_POST['targa']) and isset($_POST['matricola']) and isset($_POST['effrazioni']) and isset($_POST['luogo']) and isset($_POST['importo']))
         {
-            
-            $statement = $this->$db_connection->prepare("INSERT INTO multe VALUES (NULL,:matricola,DEFAULT,:luogo,:importo,:foto,:longitudine,:latitudine)");
-
-            $statement->execute(['matricola'=>$matricola, 'importo'=>$importo, 'luogo' => $luogo, ':foto' => $foto, ':longitudine' => $longitudine, ':latitudine' => $latitudine]);
-            
-            $idInserito =$this->$db_connection->lastInsertID();
-            echo $prova;
-
-			foreach($effrazioni as $effrazione)
-            {
+            try
+            { 
+                $statement = $this->$db_connection->prepare("INSERT INTO multe VALUES (NULL,:matricola,DEFAULT,:targa,:luogo,:importo,:foto,:longitudine,:latitudine)");
+                echo $targa;
+                $statement->execute(['matricola'=>$matricola, 'targa' => $targa, 'importo'=>$importo, 'luogo' => $luogo, ':foto' => $foto, ':longitudine' => $longitudine, ':latitudine' => $latitudine]);
+                echo "ciao 2";
+                $idInserito =$this->$db_connection->lastInsertID();
+                //echo $prova;
+                echo "ciao 3";
+			    foreach($effrazioni as $effrazione)
+                {
 					$statement = $this->$db_connection->prepare("INSERT INTO contiene VALUES(:id_Multa, :id_Effrazione)");
 					$statement->execute([':id_Multa' => $idInserito, 'id_Effrazione' => $effrazione]);
-			}
+			    }
+                echo "La multa è stata inserita";
+            }
+            catch(PDOException $e)
+            {
+                //echo "non funziona";
+                sendResponse(500, $e->getMessage());
+            }
         }
-        catch(PDOException $e)
+        else 
         {
-            //echo "non funziona";
-            sendResponse(500, $e->getMessage());
+            sendResponse(500, json_encode("Si prega di verificare che tutti i valori siano inseriti"));
         }
     }
 
@@ -87,25 +95,19 @@ class serviceApi
     {
        try
        {
-           //////////////////////////Da fare meglio più i controlli 
             $matricola = $_POST['matricola'];
 
             if("" == trim($_POST['matricola']))
             {
                 sendResponse(200, json_encode("Si prega di mettere l'ID dell'operatore"), "application/json");
-                //echo $matricola;
             }
             else{
-                //$statement = $this->$db_connection->prepare("INSERT INTO multe VALUES (NULL,:matricola,DEFAULT,:luogo,:importo,:foto,:longitudine,:latitudine)");
-                $statement = $this->$db_connection->prepare("SELECT * FROM multe WHERE ID_Operatore = :matricola");
 
+                $statement = $this->$db_connection->prepare("SELECT T1.ID_Multa, T1.Targa, T1.DataOra,T1.Luogo,T1.importo_da_pagare, GROUP_CONCAT( ' ', T3.Tipologia) AS TipologiaEffrazioni, GROUP_CONCAT(T3.ID_effrazione) AS IDEffrazioni FROM multe AS T1 INNER JOIN contiene AS T2 ON(T1.ID_Multa = T2.ID_Multa) INNER JOIN tipoeffrazione AS T3 ON(T2.ID_effrazione = T3.ID_effrazione) WHERE T1.ID_Operatore = :matricola GROUP BY T1.ID_Multa,T1.DataOra,T1.Luogo,T1.importo_da_pagare ORDER BY `T1`.`importo_da_pagare` ASC");
                 $statement->execute([':matricola' => $matricola]);
-
-                //$miaQuery = "SELECT * FROM multe WHERE ID_Operatore = $matricola";
-                //$statement = $this->$db_connection->query($miaQuery, PDO::FETCH_ASSOC);
-
+                
                 $risultati = $statement->fetchAll();
-                //echo $miaQuery;
+
                 sendResponse(200, json_encode($risultati), "application/json");
             }  
        }
@@ -117,38 +119,29 @@ class serviceApi
 
     function annullaMulta()
     {
-        //echo "Hai chiesto di annullare una multa";
-        //DELETE FROM table_name WHERE condition
-        /*
-        $statement = $this->$db_connection->prepare("SELECT * FROM multe WHERE ID_Operatore = :matricola");
-        $statement->execute([':matricola' => $matricola]);
-        */
-
-        //////////////////////////Da fare meglio più i controlli 
+        
         $multadaCanc = $_POST['id_Multa'];
-        $statement = $this->$db_connection->prepare("DELETE from multe where ID_Multa = :id");
-        //$miaQuery = "DELETE from multe where ID_Multa = $multadaCanc";
-        if("" == trim($_POST['matricola']))
+
+        if(isset($_POST['id_Multa']))
         {
             try
             {
+                
+                $statement = $this->$db_connection->prepare("DELETE from multe where ID_Multa = :id");
                 $statement->execute([':id' => $multadaCanc]);
+
                 $risultati = $statement->fetchAll();
-                //echo $miaQuery;
                 sendResponse(200, json_encode($risultati), "application/json");
             }
             catch(PDOException $e)
             {
-                sendResponse(500, $e->getMessage());
+                //sendResponse(500, $e->getMessage());
             }
         }
         else
         {
             sendResponse(200, json_encode("Si prega di mettere l'ID della multa da cancellare"), "application/json");
         }
-
-        
-
     }
 
     function richiediSanzioni()
@@ -162,13 +155,13 @@ class serviceApi
     function autenticazione()
     {
         global $autenticato;
-        try
+        global $serveAccesso;
+        if(isset($_POST['key']))
+        {
+            try
         {
             $token = $_POST['key'];
-            /*
-            $statement = $this->$db_connection->prepare("SELECT * FROM `tipoeffrazione`");
-            $statement->execute();
-            */
+            
             $statement = $this->$db_connection->prepare("SELECT Token, Durata_Token from operatore where Token = :token");
             $statement->execute([':token' => $token]);
             
@@ -182,13 +175,15 @@ class serviceApi
             //se il risultato della query è vuoto allora il token o è scaduto o non è presente nel database
             if($tokenOttenuto == "")
             {
-                sendResponse(401, json_encode("Non è stato possibile autenticarti"), "application/json");
+                //sendResponse(401, json_encode("Non è stato possibile autenticarti"), "application/json");
+                $serveAccesso = true;
             }
             else{
                 if($durataToken - time() < 0)
                 {
                     $risposta = ['esito' => "Si prega di autetnicarsi nuovamente, il token è scaduto :c"];
                     sendResponse(401, json_encode($risposta), "application/json");
+                    $serveAccesso = true;
                 }
                 else{
                     //echo "sei entrato";
@@ -198,49 +193,47 @@ class serviceApi
                      //da correggere qua la variabile lucchetto per l'autenticazione  
                      //echo $durataToken;
                 }
+
             }
         }
         catch(PDOException $e)
-        {
+            {
             sendResponse(500, $e->getMessage());
-        }
+            }
+    }
+        else{
+            sendResponse(401, json_encode("Token mancante"), "application/json");
+        }   
         return;
     }
 
     function accesso()
     {
         
-        $matr = $_POST['Matr'];
-        $pass = $_POST['Pass'];
-        //$miaQuery = "SELECT * from operatore where ID_Operatore = '$user' AND password = '$pass'";
-        //echo $miaQuery;
+        $matr = $_POST['matricola'];
+        $hashPass = hash('sha512', $_POST['password']);
+        
         
         $statement = $this->$db_connection->prepare("SELECT * from operatore where ID_Operatore = :matricola AND password = :password");
-        $statement->execute([':matricola' => $matr, ':password' => $pass]);
+        $statement->execute([':matricola' => $matr, ':password' => $hashPass]);
         
-        //$statement = $this->$db_connection->query($miaQuery, PDO::FETCH_ASSOC);  
         $risultati = $statement->fetchAll();
         $vettoreCredenzialiOttenuto = $risultati[0];
         if(isset($vettoreCredenzialiOttenuto))
         {
-           // echo "sono state trovate le credenziali";
-            $nuovoToken = hash("md5",$pass.time());
-            //echo $nuovoToken;
+            $nuovoToken = hash("md5",$hashPass.time());
 
-            //$miaQuery = "UPDATE operatore SET Token = '$nuovoToken' WHERE operatore.ID_Operatore = '$user' AND operatore.password = '$pass'";
             $statement = $this->$db_connection->prepare("UPDATE operatore SET Token = :token WHERE operatore.ID_Operatore = :matricola AND operatore.password = :password");
-            $statement->execute([':matricola' => $matr, ':password' => $pass,'token'=> $nuovoToken]);
+            $statement->execute([':matricola' => $matr, ':password' => $hashPass,'token'=> $nuovoToken]);
             $statement = $this->$db_connection->prepare("UPDATE operatore SET Durata_token = :durata WHERE operatore.ID_Operatore = :matricola AND operatore.password = :password");
-            $statement->execute([':matricola' => $matr, ':password' => $pass,'durata'=> time()+86400]);
+            $statement->execute([':matricola' => $matr, ':password' => $hashPass,'durata'=> time()+86400]);
 
-            //echo $miaQuery;
-            //$statement = $this->$db_connection->query($miaQuery);
             $risposta = ['token' => $nuovoToken, 'esito' => "Token aggiornato per 24h", ];
             sendresponse(200,json_encode($risposta),'application/json');   
         }
         else
         {
-            $risposta = [ 'esito' => "Il codice matricola o la password sono errati ", ];
+            $risposta = [ 'esito' => "Il codice matricola o la password sono errati o non presenti", ];
             sendresponse(200,json_encode($risposta),'application/json');   
         }
     }
@@ -257,16 +250,15 @@ if($autenticato == true)
 {
     switch($nome_funzione)
     {
-        case 'nuova_multa'://DA FINIRE PER BENE, MANCANO I CONTROLLI
+        case 'nuova_multa'://fatto
             $api->nuovaMulta();
         break;
     
-        case 'elenco_multe'://DA FINIRE PER BENE, MANCANO I CONTROLLI
-            //echo hash("md5","ciao".time());
+        case 'elenco_multe'://fatto
            $api->elencaMulte();
         break;
 
-        case 'annulla_multa'://quasi fatto manca cancellare più multe insieme //DA FINIRE PER BENE, MANCANO I CONTROLLI
+        case 'annulla_multa': //fatto
             $api->annullaMulta();
         break;
     
@@ -279,8 +271,8 @@ if($autenticato == true)
         break;
     }
 }
-else{
-    $api->accesso(); //DA FINIRE PER BENE, MANCANO I CONTROLLI
+else if($serveAccesso == true){
+    $api->accesso(); 
 }
-
+$api->__destruct();
 ?>
